Move helper classes from 'mojo' to 'util'
