+++ bundle-plugin/src/main/java/com/yahoo/container/plugin/mojo/AssembleTestBundleMojo.java
@@ -1,6 +1,8 @@
 // Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.container.plugin.mojo;
 
+import com.yahoo.container.plugin.util.Artifacts;
+import com.yahoo.container.plugin.util.TestBundleDependencyScopeTranslator;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
@@ -10,8 +12,8 @@ import org.codehaus.plexus.archiver.jar.JarArchiver;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import static com.yahoo.container.plugin.mojo.TestBundleUtils.archiveFile;
-import static com.yahoo.container.plugin.mojo.TestBundleUtils.manifestFile;
+import static com.yahoo.container.plugin.util.TestBundleUtils.archiveFile;
+import static com.yahoo.container.plugin.util.TestBundleUtils.manifestFile;
 
 /**
  * @author bjorncs
