+++ bundle-plugin/src/main/java/com/yahoo/container/plugin/util/TestBundleUtils.java
@@ -1,5 +1,5 @@
 // Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.container.plugin.mojo;
+package com.yahoo.container.plugin.util;
 
 import org.apache.maven.project.MavenProject;
 
@@ -10,14 +10,14 @@ import java.util.jar.JarFile;
 /**
  * @author bjorncs
  */
-class TestBundleUtils {
+public class TestBundleUtils {
     private TestBundleUtils() {}
 
-    static Path outputDirectory(MavenProject project) { return targetDirectory(project).resolve("test-bundle/"); }
+    public static Path outputDirectory(MavenProject project) { return targetDirectory(project).resolve("test-bundle/"); }
 
-    static Path manifestFile(MavenProject project) { return outputDirectory(project).resolve(JarFile.MANIFEST_NAME); }
+    public static Path manifestFile(MavenProject project) { return outputDirectory(project).resolve(JarFile.MANIFEST_NAME); }
 
-    static Path archiveFile(MavenProject project) {
+    public static Path archiveFile(MavenProject project) {
         return targetDirectory(project).resolve(project.getBuild().getFinalName() + "-tests.jar");
     }
